+++ conans/test/functional/remote/rest_api_test.py
@@ -14,7 +14,7 @@ from conans.model.ref import ConanFileReference, PackageReference
 from conans.model.version import Version
 from conans.paths import CONANFILE, CONANINFO, CONAN_MANIFEST
 from conans.server.rest.bottle_plugins.version_checker import VersionCheckerPlugin
-from conans.test.unittests.server.utils.server_launcher import TestServerLauncher
+from conans.test.utils.server_launcher import TestServerLauncher
 from conans.test.utils.test_files import hello_source_files, temp_folder
 from conans.test.utils.tools import TestBufferConanOutput
 from conans.util.files import md5, save
